AB#189099 - Split ERPFacadeConfiguration to maintain on class per file.
