+++ tests/UKHO.ERPFacade.API.FunctionalTests/Configuration/PermitWithSameKey.cs
@@ -1,0 +1,15 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Text;
+using System.Threading.Tasks;
+
+namespace UKHO.ERPFacade.API.FunctionalTests.Configuration
+{
+    public class PermitWithSameKey
+    {
+        public string Permit { get; set; }
+        public string ActiveKey { get; set; }
+        public string NextKey { get; set; }
+    }
+}
